+++ service/frontend/workflow_handler_test.go
@@ -75,6 +75,7 @@ import (
 	"go.temporal.io/server/common/rpc/interceptor"
 	"go.temporal.io/server/common/searchattribute"
 	"go.temporal.io/server/common/tasktoken"
+	"go.temporal.io/server/service/history/api"
 	"go.temporal.io/server/service/history/tests"
 	"go.temporal.io/server/service/worker/batcher"
 	"go.temporal.io/server/service/worker/scheduler"
@@ -2081,7 +2082,7 @@ func (s *WorkflowHandlerSuite) TestVerifyHistoryIsComplete() {
 	}
 
 	for i, tc := range testCases {
-		err := persistence.VerifyHistoryIsComplete(tc.events, tc.firstEventID, tc.lastEventID, tc.isFirstPage, tc.isLastPage, tc.pageSize)
+		err := api.VerifyHistoryIsComplete(tc.events, tc.firstEventID, tc.lastEventID, tc.isFirstPage, tc.isLastPage, tc.pageSize)
 		if tc.isResultErr {
 			s.Error(err, "testcase %v failed", i)
 		} else {
@@ -3325,84 +3326,5 @@ func (s *WorkflowHandlerSuite) TestShutdownWorker() {
 	})
 	if err != nil {
 		s.Fail("ShutdownWorker failed:", err)
-	}
-}
-
-// HistoryService returns a different GetWorkflowExecutionHistoryResponse GetWorkflowExecutionHistoryResponseWithRaw to
-// WorkflowHandler. Since HistoryClient is defined with the response type GetWorkflowExecutionHistoryResponse, grpc
-// will deserialize this message to GetWorkflowExecutionHistoryResponse. This is done to avoid the extra CPU usage in
-// history service to deserialize event blobs to []*HistoryEvent. This test ensures that
-// GetWorkflowExecutionHistoryResponseWithRaw is correctly deserialized to GetWorkflowExecutionHistoryResponse.
-func (s *WorkflowHandlerSuite) TestGetWorkflowExecutionHistoryResponseWithRawHistoryEvents() {
-	// Create history events and batches
-	fullHistory := &historypb.History{
-		Events: []*historypb.HistoryEvent{
-			{
-				EventId:   1,
-				EventType: enumspb.EVENT_TYPE_WORKFLOW_EXECUTION_STARTED,
-				Version:   100,
-			},
-			{
-				EventId:   2,
-				EventType: enumspb.EVENT_TYPE_WORKFLOW_TASK_SCHEDULED,
-				Version:   101,
-			},
-			{
-				EventId:   3,
-				EventType: enumspb.EVENT_TYPE_ACTIVITY_TASK_COMPLETED,
-				Version:   102,
-			},
-			{
-				EventId:   4,
-				EventType: enumspb.EVENT_TYPE_TIMER_FIRED,
-				Version:   103,
-			},
-		},
-	}
-
-	batch1 := &historypb.History{
-		Events: fullHistory.Events[:2],
-	}
-
-	batch2 := &historypb.History{
-		Events: fullHistory.Events[2:],
-	}
-
-	// Marshal each batch
-	rawHistory1, err := batch1.Marshal()
-	s.Require().NoError(err)
-
-	db1 := &commonpb.DataBlob{
-		EncodingType: enumspb.ENCODING_TYPE_PROTO3,
-		Data:         rawHistory1,
-	}
-
-	rawHistory2, err := batch2.Marshal()
-	s.Require().NoError(err)
-
-	db2 := &commonpb.DataBlob{
-		EncodingType: enumspb.ENCODING_TYPE_PROTO3,
-		Data:         rawHistory2,
-	}
-
-	// Create a GetWorkflowExecutionHistoryResponse with raw history blobs. This should be wire compatible with
-	// workflowservice.GetWorkflowExecutionHistoryResponse which has repeated HistoryEvent field instead of repeated bytes.
-	rawResp := &historyspb.GetWorkflowExecutionHistoryResponse{
-		History: [][]byte{db1.Data, db2.Data},
-	}
-
-	serializedRawResp, err := rawResp.Marshal()
-	s.Require().NoError(err)
-
-	resp := &workflowservice.GetWorkflowExecutionHistoryResponse{}
-	err = resp.Unmarshal(serializedRawResp)
-	s.Require().NoError(err)
-
-	// Verify resp has same list of history events as fullHistory
-	for i, event := range resp.History.Events {
-		s.Equal(fullHistory.Events[i].EventId, event.EventId)
-		s.Equal(fullHistory.Events[i].Version, event.Version)
-		s.Equal(fullHistory.Events[i].EventType, event.EventType)
-		s.Nil(event.Attributes)
 	}
 }
